[#3954] Create a separate module for the caching code
